[move/#5] 컴포넌트 폴더
